[improvements] - [`nexial.inspectOnPause`]: now support "end immediate" via `:end`.
[expression]
- [`nexial.expression.web.alwaysNew`]: *NEW* System variable to ensure each use of WEB expression clear off outcome from the WEB operations from previous use.

+++ src/main/kotlin/org/nexial/core/plugins/db/DBDriverHelper.kt
@@ -167,7 +167,7 @@ abstract class DBDriverHelper protected constructor(protected var context: Execu
 
     companion object {
 
-        const val DRIVER_MIN_SIZE: Long = 1024 * 50
+        const val DRIVER_MIN_SIZE: Long = (1024 * 50).toLong()
 
         @JvmStatic
         @Throws(IOException::class)
